Update membership type to be a drop down list data type.

+++ website/usercontrols/cFront/RegisterMemberComplete.ascx.cs
@@ -107,7 +107,7 @@ public partial class usercontrols_cFront_RegisterMemberComplete : System.Web.UI.
 
 	private void SetMembershipOptions(IDictionary<String, object> currentmemdata, MembershipOptions membershipOptions)
 	{
-		currentmemdata[MemberProperty.membershipType] = membershipOptions.MembershipType.ToString();
+		currentmemdata[MemberProperty.membershipType] = ((int) membershipOptions.MembershipType).ToString();
 		currentmemdata[MemberProperty.OpenWaterIndemnityAcceptance] = membershipOptions.OpenWaterIndemnityAcceptance;
 		currentmemdata[MemberProperty.swimSubsJanToJune] = membershipOptions.SwimSubsJanToJune;
 		currentmemdata[MemberProperty.SwimSubsJulyToDec] = membershipOptions.SwimSubsJulyToDec;
